♻️ refactor: 회원 정보와 로그인 클래스 분리 [DRKKO-55]
